ridedott-pmc-hackathon/pyproject.toml
[tool.poetry]
name = "ridedott-pmc-hackathon"
version = "0.1.0"
description = ""
authors = ["Alexvonrass <alexvonrass@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.2"
vertexai = "^1.48.0"
weaviate-client = "^4.5.6"
google-cloud-aiplatform = "^1.48.0"
google-cloud-vision = "^3.7.2"
google-cloud-speech = "^2.26.0"
pytest = "^8.1.1"
pre-commit = "^3.7.0"
diagrams = "^0.23.4"
dash = "^2.16.1"
dash-bootstrap-components = "^1.6.0"
dash-core-components = "^2.0.0"
dash-recording-components = "^1.0.3"
soundfile = "^0.12.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 79
exclude = [".git", "*migrations*", ".idea", ".pytest_cache", ".tmp", ".venv", "venv"]
ignore = ["E501", "W503"]
extend-ignore = ["E203"]
max-complexity = 15

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
ensure_newline_before_comments = true
